Reorganize users tests
